+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # (MODX)EvolutionCMS.snippets.ddTypograph changelog
 
 
+## Version 2.7 (2024-07-07)
+* \+ You can turn off typography for a text fragment aslo by using the `<span class="notg"></span>` tag.
+
+
 ## Version 2.6.1 (2023-08-23)
 * \* TinyMCE empty lines (`<p>&nbsp;</p>`) are no longer modified.
 
@@ -19,7 +23,7 @@
 * \* Attention! PHP >= 5.6 is required.
 * \* Attention! (MODX)EvolutionCMS.libraries.ddTools >= 0.48.1 is required.
 * \+ You can just call `\DDTools\Snippet::runSnippet` to run the snippet without DB and eval (see README → Examples).
-* \+ `\ddDomainComparator\Snippet`: The new class. All snippet code was moved here.
+* \+ `\ddTypograph\Snippet`: The new class. All snippet code was moved here.
 * \+ README → Links.
 * \* Composer.json:
 	* \+ `homepage`